pyproject.toml
[project]
name = "todo-list-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "fastapi-limiter>=0.1.6",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.40.0",
]
